news_extractor-main/requirements.txt
azure-core~=1.32.0
protobuf~=5.29.3
certifi~=2025.1.31
pymongo~=4.11.2
openai~=1.65.3
langchain-openai~=0.3.7
beautifulsoup4